+++ src/react/calculator/src/App.css
@@ -1,44 +1,21 @@
-#members {
-  width: 300px;
+.App {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
 }
 
-
-
-.input-group {
-  width: 300px;
-
+table {
+    border: 2px solid forestgreen;
+    width: 800px;
+    height: 200px;
 }
 
-#select {
-  width: 300px;
+th {
+    border-bottom: 1px solid black;
 }
 
-
-.checkbox {
-  /* style the dropdown */
-  width: 200px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-.checkbox div {
-  /* style the options in the dropdown */
-  display: flex;
-  align-items: center;
-  padding: 5px 10px;
-  cursor: pointer;
-}
-
-.checkbox div:hover {
-  /* style the options when they are hovered over */
-  background-color: #eee;
-}
-
-.checkbox input[type="checkbox"] {
-  /* style the checkboxes */
-  margin-right: 10px;
-}
-
-.checkbox label {
-  /* style the labels */
+td {
+    text-align: center;
 }
